+++ src/main/java/com/herdlicka/igneousmachines/block/IgneousMinerBlock.java
@@ -168,9 +168,9 @@ public class IgneousMinerBlock extends BlockWithEntity {
         boolean bl = world.isReceivingRedstonePower(pos) || world.isReceivingRedstonePower(pos.up());
         boolean bl2 = state.get(TRIGGERED);
         if (bl && !bl2) {
-            world.setBlockState(pos, state.with(TRIGGERED, true), Block.NO_REDRAW);
+            world.setBlockState(pos, state.with(TRIGGERED, true), Block.NOTIFY_LISTENERS);
         } else if (!bl && bl2) {
-            world.setBlockState(pos, state.with(TRIGGERED, false), Block.NO_REDRAW);
+            world.setBlockState(pos, state.with(TRIGGERED, false), Block.NOTIFY_LISTENERS);
         }
     }
 }